+++ src/AddCard/components/AddCard.tsx
@@ -1,0 +1,15 @@
+import AddCardForm from "./AddCardForm/components/AddCardForm";
+import useAddCard from "../hooks/useAddCard";
+
+function AddCard() {
+  const { addCardState, previewState } = useAddCard();
+
+  return (
+    <>
+      <div>Preview</div>
+      <AddCardForm addCardState={addCardState} />
+    </>
+  );
+}
+
+export default AddCard;
